refactor: keep the flat camera maths in MercatorTransform and invert _calcMatrices for the sphere camera options
